requirements.txt
flask
requests
python-dotenv
pandas
apscheduler
matplotlib
